refactor: app structure improvements - moved leftMenu and rightManu state to ui store - simplified App.tsx by moving panes logic to new PaneManager component
