+++ src/components/organisms/index.ts
@@ -8,3 +8,4 @@ export {default as PageFooter} from './PageFooter';
 export {default as PageHeader} from './PageHeader';
 export {default as SettingsDrawer} from './SettingsDrawer';
 export {default as HotKeysHandler} from './HotKeysHandler';
+export {default as PaneManager} from './PaneManager';
