+++ src/models/ui.ts
@@ -1,4 +1,12 @@
 export type UiState = {
   settingsOpened: boolean;
   isFolderLoading: boolean;
+  leftMenu: {
+    selection: string;
+    isActive: boolean;
+  };
+  rightMenu: {
+    selection?: string;
+    isActive: boolean;
+  };
 };
