+++ src/redux/reducers/ui.ts
@@ -1,4 +1,4 @@
-import {createSlice, Draft} from '@reduxjs/toolkit';
+import {createSlice, Draft, PayloadAction} from '@reduxjs/toolkit';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
 import {UiState} from '@models/ui';
 import initialState from '@redux/initialState';
@@ -9,6 +9,18 @@ export const uiSlice = createSlice({
   reducers: {
     toggleSettings: (state: Draft<UiState>) => {
       state.settingsOpened = !state.settingsOpened;
+    },
+    toggleLeftMenu: (state: Draft<UiState>) => {
+      state.leftMenu.isActive = !state.leftMenu.isActive;
+    },
+    setLeftMenuSelection: (state: Draft<UiState>, action: PayloadAction<string>) => {
+      state.leftMenu.selection = action.payload;
+    },
+    toggleRightMenu: (state: Draft<UiState>) => {
+      state.rightMenu.isActive = !state.rightMenu.isActive;
+    },
+    setRightMenuSelection: (state: Draft<UiState>, action: PayloadAction<string>) => {
+      state.rightMenu.selection = action.payload;
     },
   },
   extraReducers: builder => {
@@ -25,5 +37,6 @@ export const uiSlice = createSlice({
   },
 });
 
-export const {toggleSettings} = uiSlice.actions;
+export const {toggleSettings, toggleLeftMenu, toggleRightMenu, setLeftMenuSelection, setRightMenuSelection} =
+  uiSlice.actions;
 export default uiSlice.reducer;
